feat: Added multiple base files
[saveData]
No working code added as of now.
Simply made for clarity and to know
what I want and how it will look in the future.

[calculateData]
Got this from the other branch.
It calculates the average of a day and
can sort a list.

[Comments]
I added in all the .py files what they
are supposed to do and what their role
is in the whole project.
This wil clarify to any user checking the
code which file does which.