feat: added calculation updates and other minor changes
[CalculateData]
Added calculate lowest price, can possibly be used
for calculating how fast the vehicle or device should
charge.
Removed sortList due to it not working as intended.

[SaveData]
Added a basic lay-out to the function with a dictionary
and some empty functions which will be configured in the
future.

[Main]
Did some testing with the new functions. No big changes
were made here even though it says it's modified.