feat: added local database support
[SQLite]
Added a function to locally make a SQLite database
and save data there, currently used to gather some
data over a short period of time.
Currently, the date and prices per hour (IN KWH) are
saved in the database.

[NOT YET INCLUDED]
I want to add another thing but this is completely local
to my own house, my solar panels's output. This could
be handy to include for data for myself.