fix: fixed data.db saving format
[database ID]
Firstly the ID of each cell was bugged because I did the
math wrong.
I changed it to hard coded numbers, this is not optimal
and SHOULD be changed. to the previous version but *10 it all.

[database prices]
I saved the prices so that you cannot see which number is actually
which number. I made a small script which fixes the strings and
makes it back to a list again, this can be helpfull to decode
the database if this is further needed.